+++ src/writers/writer.hh
@@ -6,6 +6,7 @@
 
 #include "common/dimension.hh"
 #include "common/thread.pool.hh"
+#include "common/s3.connection.hh"
 #include "blosc.compressor.hh"
 #include "file.sink.hh"
 
@@ -40,7 +41,8 @@ struct Writer
   public:
     Writer() = delete;
     Writer(const WriterConfig& config,
-           std::shared_ptr<common::ThreadPool> thread_pool);
+           std::shared_ptr<common::ThreadPool> thread_pool,
+           std::shared_ptr<common::S3ConnectionPool> connection_pool);
 
     virtual ~Writer() noexcept = default;
 
@@ -70,6 +72,8 @@ struct Writer
     uint32_t frames_written_;
     uint32_t append_chunk_index_;
     bool is_finalizing_;
+
+    std::shared_ptr<common::S3ConnectionPool> connection_pool_;
 
     void make_buffers_() noexcept;
     size_t write_frame_to_chunks_(const uint8_t* buf, size_t buf_size);